+++ build_tools/aws-sdk-code-generator/lib/aws-sdk-code-generator/views/client_class.rb
@@ -80,6 +80,11 @@ module AwsSdkCodeGenerator
         Waiter.markdown_table(@waiters)
       end
 
+      # @return [Boolean]
+      def authorizer?
+        @custom
+      end
+
     end
   end
 end